Fix opencv error in cv2.resize()
Apprarently at some point the image is not an integer anymore which is then converted to -1 by opencv.
Thus, I assure the datatype to be int of the input image for reshape

+++ pysot/tools/knn_classifier/knn.py
@@ -12,7 +12,7 @@ size = {"width": 450, "height": 450}
 
 
 def resize(image, width=None, height=None, inter=cv2.INTER_AREA):
-    return cv2.resize(image, (width, height), interpolation=inter)
+    return cv2.resize(image.astype("uint8"), (width, height), interpolation=inter)
 
 
 def resize_images(images):
